gba: Change GBT format to fix arpeggio
If the frequency comes after the effect, the arpeggio command handler
doesn't know the base frequency index when calculating the frequency
indices of the other arpeggio steps.

This patch moves the frequency to the first thing in a step, right after
the volume, so that when the effect is handled the desired base
frequency has been read.