Use old sdcc calling convention when new GBDK-2020 present
We've had a few users trying to use gbt-player with newer GBDK-2020 releases. In these releases SDCC uses the new calling convention __sdcccall(1) by default. This causes problems with some gbt-player calls which expect the old calling convention.

This PR changes the header to specify that the old calling convention __sdcccall(0) should be used when built with newer SDCC versions. If an older SDCC is used then the calling convention request will be ignored, allowing it to be compatible in both scenarios.